requirements.txt
annotated-types==0.7.0
beautifulsoup4==4.12.2
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
Cython==3.0.8
et_xmlfile==2.0.0
ezdxf==1.4.0
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
fonttools==4.56.0
geographiclib==1.52
geopy==2.4.1
greenlet==3.2.2
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
labelImg==1.8.6
Levenshtein==0.27.1
lxml==5.1.0
mapbox-vector-tile==2.1.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mercantile==1.2.1
mpmath==1.3.0
numpy==2.3.3
opencv-python==4.11.0.86
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
pillow==11.2.1
playwright==1.52.0
protobuf==5.29.4
psycopg2-binary==2.9.10
pyclipper==1.3.0.post6
pycocotools==2.0.7
pydantic_core==2.33.2
pydub==0.25.1
pyee==13.0.0
PyOpenGL==3.1.7
PyOpenGL-accelerate==3.1.7
pyparsing==3.2.3
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyshp==2.3.1
python-dateutil==2.9.0.post0
python-dotenv==0.19.0
python-multipart==0.0.20
pytz==2025.2
RapidFuzz==3.13.0
requests==2.31.0
shapely==2.0.7
shellingham==1.5.4
simplekml==1.3.6
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.43
sympy==1.13.1
tomlkit==0.12.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==1.26.20
uvicorn==0.34.3
websockets==11.0.3
Werkzeug==3.1.3
